add/remove-shell: also add/remove shells real path
In the case of usr merged systems we want the shells to also
be listed with their real paths, eg. /usr/bin/bash, in addition
to their specified path in /etc/shells

(While at it also fix mixed tab/spaces indentation of help texts
and trailing whitespace.)

The mixture of dirname/realpath/basename is to avoid expanding
the filename, only the directory part of the path.
This is to avoid /bin/sh being expanded to /usr/bin/dash
rather than /usr/bin/sh. Maybe we want to add both in
such a case?

Closes: #817168