implement proper quoting on add-shell and remove-shell
The quoting of $file, $tmpfile and $otmpfile is not necessary yet as
their values are known to not contain spaces. Once adding DPKG_ROOT
support, lack of quoting will become problematic though. Lack of quoting
is a theoretical issue for arguments. In any case, this commit makes
shellcheck/0.7.1 stop complaining about the scripts.